+++ src/main/java/com/mysql/MySqlReservationDao.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManagerFactory;
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -103,7 +104,7 @@ public class MySqlReservationDao implements ReservationDao {
     }
 
     @Override
-    public Reservation getReservationById(int id) throws DaoException {
+    public Reservation getReservationById(BigInteger id) throws DaoException {
         try {
             session = sessionFactory.openSession();
             session.beginTransaction();
